Implement Colossal Blood Chest repairing algorithm This machine becomes more efficient when more items are repaired simultaneously. An efficiency bar can be seen in the machine gui.

+++ src/main/java/evilcraft/core/tileentity/TickingTankInventoryTileEntity.java
@@ -86,7 +86,7 @@ public abstract class TickingTankInventoryTileEntity<T extends TankInventoryTile
                 NBTTagCompound tag = tickerList.getCompoundTagAt(i);
                 if(tag != null) {
                     ticker.setTick(tag.getInteger("tick"));
-                    ticker.setRequiredTicks(tag.getInteger("requiredTicks"));
+                    ticker.setRequiredTicks(tag.getFloat("requiredTicks"));
                 }
             }
         }
@@ -100,7 +100,7 @@ public abstract class TickingTankInventoryTileEntity<T extends TankInventoryTile
         for(TickComponent<T, ITickAction<T>> ticker : tickers) {
             NBTTagCompound tag = new NBTTagCompound();
             tag.setInteger("tick", ticker.getTick());
-            tag.setInteger("requiredTicks", ticker.getRequiredTicks());
+            tag.setFloat("requiredTicks", ticker.getRequiredTicks());
             tickerList.appendTag(tag);
         }
         data.setTag("tickers", tickerList);
